Trainers: split tasks into separate files
