--- torchgeo/trainers/chesapeake.py
@@ -1,283 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-"""Trainers for the Chesapeake datasets."""
-
-from typing import Any, Dict, cast
-
-import matplotlib
-import matplotlib.pyplot as plt
-import numpy as np
-import segmentation_models_pytorch as smp
-import torch
-import torch.nn as nn
-from pytorch_lightning.core.lightning import LightningModule
-from torch import Tensor
-from torch.optim.lr_scheduler import ReduceLROnPlateau
-from torch.utils.tensorboard import SummaryWriter  # type: ignore[attr-defined]
-from torchmetrics import Accuracy, IoU, MetricCollection
-
-from ..datasets import Chesapeake7
-from ..models import FCN
-
-# TODO: move the color maps to a dataset object
-CMAP_7 = matplotlib.colors.ListedColormap(
-    [np.array(Chesapeake7.cmap[i]) / 255.0 for i in range(7)]
-)
-CMAP_5 = matplotlib.colors.ListedColormap(
-    np.array(
-        [
-            (0, 0, 0, 0),
-            (0, 197, 255, 255),
-            (38, 115, 0, 255),
-            (163, 255, 115, 255),
-            (156, 156, 156, 255),
-        ]
-    )
-    / 255.0
-)
-
-
-class ChesapeakeCVPRSegmentationTask(LightningModule):
-    """LightningModule for training models on the Chesapeake CVPR Land Cover dataset.
-
-    This allows using arbitrary models and losses from the
-    ``pytorch_segmentation_models`` package.
-    """
-
-    def config_task(self) -> None:
-        """Configures the task based on kwargs parameters passed to the constructor."""
-        if self.hparams["class_set"] not in [5, 7]:
-            raise ValueError("'class_set' must be either 5 or 7")
-        num_classes = self.hparams["class_set"]
-        classes = range(1, self.hparams["class_set"])
-
-        if self.hparams["segmentation_model"] == "unet":
-            self.model = smp.Unet(
-                encoder_name=self.hparams["encoder_name"],
-                encoder_weights=self.hparams["encoder_weights"],
-                in_channels=4,
-                classes=num_classes,
-            )
-        elif self.hparams["segmentation_model"] == "deeplabv3+":
-            self.model = smp.DeepLabV3Plus(
-                encoder_name=self.hparams["encoder_name"],
-                encoder_weights=self.hparams["encoder_weights"],
-                in_channels=4,
-                classes=num_classes,
-            )
-        elif self.hparams["segmentation_model"] == "fcn":
-            self.model = FCN(in_channels=4, classes=num_classes, num_filters=256)
-        else:
-            raise ValueError(
-                f"Model type '{self.hparams['segmentation_model']}' is not valid."
-            )
-
-        if self.hparams["loss"] == "ce":
-            self.loss = nn.CrossEntropyLoss(  # type: ignore[attr-defined]
-                ignore_index=0
-            )
-        elif self.hparams["loss"] == "jaccard":
-            self.loss = smp.losses.JaccardLoss(mode="multiclass", classes=classes)
-        elif self.hparams["loss"] == "focal":
-            self.loss = smp.losses.FocalLoss(
-                "multiclass", ignore_index=0, normalized=True
-            )
-        else:
-            raise ValueError(f"Loss type '{self.hparams['loss']}' is not valid.")
-
-    def __init__(self, **kwargs: Any) -> None:
-        """Initialize the LightningModule with a model and loss function.
-
-        Keyword Args:
-            segmentation_model: Name of the segmentation model type to use
-            encoder_name: Name of the encoder model backbone to use
-            encoder_weights: None or "imagenet" to use imagenet pretrained weights in
-                the encoder model
-            loss: Name of the loss function
-
-        Raises:
-            ValueError: if kwargs arguments are invalid
-        """
-        super().__init__()
-        self.save_hyperparameters()  # creates `self.hparams` from kwargs
-
-        self.config_task()
-
-        self.train_metrics = MetricCollection(
-            [
-                Accuracy(num_classes=self.hparams["class_set"], ignore_index=0),
-                IoU(num_classes=self.hparams["class_set"], ignore_index=0),
-            ],
-            prefix="train_",
-        )
-        self.val_metrics = self.train_metrics.clone(prefix="val_")
-        self.test_metrics = self.train_metrics.clone(prefix="test_")
-
-    def forward(self, x: Tensor) -> Any:  # type: ignore[override]
-        """Forward pass of the model.
-
-        Args:
-            x: tensor of data to run through the model
-
-        Returns:
-            output from the model
-        """
-        return self.model(x)
-
-    def training_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> Tensor:
-        """Training step - reports average accuracy and average IoU.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-
-        Returns:
-            training loss
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the train step logs every `log_every_n_steps` steps where
-        # `log_every_n_steps` is a parameter to the `Trainer` object
-        self.log("train_loss", loss, on_step=True, on_epoch=False)
-        self.train_metrics(y_hat_hard, y)
-
-        return cast(Tensor, loss)
-
-    def training_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level training metrics.
-
-        Args:
-            outputs: list of items returned by training_step
-        """
-        self.log_dict(self.train_metrics.compute())
-        self.train_metrics.reset()
-
-    def validation_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Validation step - reports average accuracy and average IoU.
-
-        Logs the first 10 validation samples to tensorboard as images with 3 subplots
-        showing the image, mask, and predictions.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        self.log("val_loss", loss, on_step=False, on_epoch=True)
-        self.val_metrics(y_hat_hard, y)
-
-        if batch_idx < 10:
-            cmap = None
-            if self.hparams["class_set"] == 5:
-                cmap = CMAP_5
-            else:
-                cmap = CMAP_7
-            # Render the image, ground truth mask, and predicted mask for the first
-            # image in the batch
-            img = np.rollaxis(  # convert image to channels last format
-                batch["image"][0].cpu().numpy(), 0, 3
-            )
-            mask = batch["mask"][0].cpu().numpy()
-            pred = y_hat_hard[0].cpu().numpy()
-            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
-            axs[0].imshow(img[:, :, :3])
-            axs[0].axis("off")
-            axs[1].imshow(
-                mask,
-                vmin=0,
-                vmax=self.hparams["class_set"] - 1,
-                cmap=cmap,
-                interpolation="none",
-            )
-            axs[1].axis("off")
-            axs[2].imshow(
-                pred,
-                vmin=0,
-                vmax=self.hparams["class_set"] - 1,
-                cmap=cmap,
-                interpolation="none",
-            )
-            axs[2].axis("off")
-            plt.tight_layout()
-
-            # the SummaryWriter is a tensorboard object, see:
-            # https://pytorch.org/docs/stable/tensorboard.html#
-            summary_writer: SummaryWriter = self.logger.experiment
-            summary_writer.add_figure(
-                f"image/{batch_idx}", fig, global_step=self.global_step
-            )
-            plt.close()
-
-    def validation_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level validation metrics.
-
-        Args:
-            outputs: list of items returned by validation_step
-        """
-        self.log_dict(self.val_metrics.compute())
-        self.val_metrics.reset()
-
-    def test_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Test step identical to the validation step.
-
-        Args:
-            batch: Current batch
-            batch_idx: Index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the test and validation steps only log per *epoch*
-        self.log("test_loss", loss, on_step=False, on_epoch=True)
-        self.test_metrics(y_hat_hard, y)
-
-    def test_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level test metrics.
-
-        Args:
-            outputs: list of items returned by test_step
-        """
-        self.log_dict(self.test_metrics.compute())
-        self.test_metrics.reset()
-
-    def configure_optimizers(self) -> Dict[str, Any]:
-        """Initialize the optimizer and learning rate scheduler.
-
-        Returns:
-            a "lr dict" according to the pytorch lightning documentation --
-            https://pytorch-lightning.readthedocs.io/en/latest/common/lightning_module.html#configure-optimizers
-        """
-        optimizer = torch.optim.Adam(
-            self.model.parameters(), lr=self.hparams["learning_rate"]
-        )
-        return {
-            "optimizer": optimizer,
-            "lr_scheduler": {
-                "scheduler": ReduceLROnPlateau(
-                    optimizer, patience=self.hparams["learning_rate_schedule_patience"]
-                ),
-                "monitor": "val_loss",
-            },
-        }
