--- torchgeo/trainers/naipchesapeake.py
@@ -1,251 +1,0 @@
-# Copyright (c) Microsoft Corporation. All rights reserved.
-# Licensed under the MIT License.
-
-"""NAIP + Chesapeake trainer."""
-
-from typing import Any, Dict, cast
-
-import matplotlib.pyplot as plt
-import numpy as np
-import pytorch_lightning as pl
-import segmentation_models_pytorch as smp
-import torch
-import torch.nn as nn
-from torch import Tensor
-from torch.optim.lr_scheduler import ReduceLROnPlateau
-from torch.utils.data import DataLoader
-from torch.utils.tensorboard import SummaryWriter  # type: ignore[attr-defined]
-from torchmetrics import Accuracy, IoU
-
-from ..models import FCN
-
-# https://github.com/pytorch/pytorch/issues/60979
-# https://github.com/pytorch/pytorch/pull/61045
-DataLoader.__module__ = "torch.utils.data"
-
-
-class NAIPChesapeakeSegmentationTask(pl.LightningModule):
-    """LightningModule for training models on the NAIP and Chesapeake datasets.
-
-    This allows using arbitrary models and losses from the
-    ``pytorch_segmentation_models`` package.
-    """
-
-    in_channels = 4
-    classes = 13
-    # TODO: tune this hyperparam
-    num_filters = 64
-
-    def config_task(self, kwargs: Any) -> None:
-        """Configures the task based on kwargs parameters."""
-        if kwargs["segmentation_model"] == "unet":
-            self.model = smp.Unet(
-                encoder_name=kwargs["encoder_name"],
-                encoder_weights=kwargs["encoder_weights"],
-                in_channels=self.in_channels,
-                classes=self.classes,
-            )
-        elif kwargs["segmentation_model"] == "deeplabv3+":
-            self.model = smp.DeepLabV3Plus(
-                encoder_name=kwargs["encoder_name"],
-                encoder_weights=kwargs["encoder_weights"],
-                encoder_output_stride=kwargs["encoder_output_stride"],
-                in_channels=self.in_channels,
-                classes=self.classes,
-            )
-        elif kwargs["segmentation_model"] == "fcn":
-            self.model = FCN(self.in_channels, self.classes, self.num_filters)
-        else:
-            raise ValueError(
-                f"Model type '{kwargs['segmentation_model']}' is not valid."
-            )
-
-        if kwargs["loss"] == "ce":
-            self.loss = nn.CrossEntropyLoss()  # type: ignore[attr-defined]
-        elif kwargs["loss"] == "jaccard":
-            self.loss = smp.losses.JaccardLoss(mode="multiclass")
-        else:
-            raise ValueError(f"Loss type '{kwargs['loss']}' is not valid.")
-
-    def __init__(self, **kwargs: Any) -> None:
-        """Initialize the LightningModule with a model and loss function.
-
-        Keyword Args:
-            segmentation_model: Name of the segmentation model type to use
-            encoder_name: Name of the encoder model backbone to use
-            encoder_weights: None or "imagenet" to use imagenet pretrained weights in
-                the encoder model
-            encoder_output_stride: The output stride parameter in DeepLabV3+ models
-            loss: Name of the loss function
-        """
-        super().__init__()
-        self.save_hyperparameters()  # creates `self.hparams` from kwargs
-
-        self.config_task(kwargs)
-
-        self.train_accuracy = Accuracy()
-        self.val_accuracy = Accuracy()
-        self.test_accuracy = Accuracy()
-
-        self.train_iou = IoU(num_classes=self.classes)
-        self.val_iou = IoU(num_classes=self.classes)
-        self.test_iou = IoU(num_classes=self.classes)
-
-    def forward(self, x: Tensor) -> Any:  # type: ignore[override]
-        """Forward pass of the model.
-
-        Args:
-            x: input image
-
-        Returns:
-            prediction
-        """
-        return self.model(x)
-
-    def training_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> Tensor:
-        """Training step - reports average accuracy and average IoU.
-
-        Args:
-            batch: current batch
-            batch_idx: index of current batch
-
-        Returns:
-            training loss
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the train step logs every `log_every_n_steps` steps where
-        # `log_every_n_steps` is a parameter to the `Trainer` object
-        self.log("train_loss", loss, on_step=True, on_epoch=False)
-        self.train_accuracy(y_hat_hard, y)
-        self.train_iou(y_hat_hard, y)
-
-        return cast(Tensor, loss)
-
-    def training_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level training metrics.
-
-        Args:
-            outputs: list of items returned by training_step
-        """
-        self.log("train_acc", self.train_accuracy.compute())
-        self.log("train_iou", self.train_iou.compute())
-        self.train_accuracy.reset()
-        self.train_iou.reset()
-
-    def validation_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Validation step - reports average accuracy and average IoU.
-
-        Args:
-            batch: current batch
-            batch_idx: index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the test and validation steps only log per *epoch*
-        self.log("val_loss", loss)
-        self.val_accuracy(y_hat_hard, y)
-        self.val_iou(y_hat_hard, y)
-
-        if batch_idx < 10:
-            # Render the image, ground truth mask, and predicted mask for the first
-            # image in the batch
-            img = np.rollaxis(  # convert image to channels last format
-                batch["image"][0].cpu().numpy(), 0, 3
-            )
-            mask = batch["mask"][0].cpu().numpy()
-            pred = y_hat_hard[0].cpu().numpy()
-            fig, axs = plt.subplots(1, 3, figsize=(12, 4))
-            axs[0].imshow(img)
-            axs[0].axis("off")
-            axs[1].imshow(mask, vmin=0, vmax=4)
-            axs[1].axis("off")
-            axs[2].imshow(pred, vmin=0, vmax=4)
-            axs[2].axis("off")
-
-            # the SummaryWriter is a tensorboard object, see:
-            # https://pytorch.org/docs/stable/tensorboard.html#
-            summary_writer: SummaryWriter = self.logger.experiment
-            summary_writer.add_figure(
-                f"image/{batch_idx}", fig, global_step=self.global_step
-            )
-
-            plt.close()
-
-    def validation_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level validation metrics.
-
-        Args:
-            outputs: list of items returned by validation_step
-        """
-        self.log("val_acc", self.val_accuracy.compute())
-        self.log("val_iou", self.val_iou.compute())
-        self.val_accuracy.reset()
-        self.val_iou.reset()
-
-    def test_step(  # type: ignore[override]
-        self, batch: Dict[str, Any], batch_idx: int
-    ) -> None:
-        """Test step identical to the validation step.
-
-        Args:
-            batch: current batch
-            batch_idx: index of current batch
-        """
-        x = batch["image"]
-        y = batch["mask"]
-        y_hat = self.forward(x)
-        y_hat_hard = y_hat.argmax(dim=1)
-
-        loss = self.loss(y_hat, y)
-
-        # by default, the test and validation steps only log per *epoch*
-        self.log("test_loss", loss)
-        self.test_accuracy(y_hat_hard, y)
-        self.test_iou(y_hat_hard, y)
-
-    def test_epoch_end(self, outputs: Any) -> None:
-        """Logs epoch level test metrics.
-
-        Args:
-            outputs: list of items returned by test_step
-        """
-        self.log("test_acc", self.test_accuracy.compute())
-        self.log("test_iou", self.test_iou.compute())
-        self.test_accuracy.reset()
-        self.test_iou.reset()
-
-    def configure_optimizers(self) -> Dict[str, Any]:
-        """Initialize the optimizer and learning rate scheduler.
-
-        Returns:
-            a "lr dict" according to the pytorch lightning documentation --
-            https://pytorch-lightning.readthedocs.io/en/latest/common/lightning_module.html#configure-optimizers
-        """
-        optimizer = torch.optim.AdamW(
-            self.model.parameters(), lr=self.hparams["learning_rate"]
-        )
-        return {
-            "optimizer": optimizer,
-            "lr_scheduler": {
-                "scheduler": ReduceLROnPlateau(
-                    optimizer, patience=self.hparams["learning_rate_schedule_patience"]
-                ),
-                "monitor": "val_loss",
-                "verbose": True,
-            },
-        }
